cleanup public surface
